Test unitarios probados y reestructuración de código
