+++ requirements.txt
@@ -1,282 +1,109 @@
-absl-py==1.2.0
-accelerate==0.13.1
-aiodns==3.0.0
-aiohttp==3.8.3
-aiosignal==1.2.0
-api-display-purposes==0.0.3
-appdirs==1.4.4
-appnope==0.1.2
-APScheduler==3.6.3
-argon2-cffi==21.1.0
+absl-py==2.1.0
+aiohttp==3.9.5
+aiosignal==1.3.1
 astunparse==1.6.3
-async-timeout==4.0.2
-attrs==22.1.0
-backcall==0.2.0
-backoff==1.11.1
-beautifulsoup4==4.9.3
-bleach==4.1.0
-boltons==23.0.0
-Brotli==1.0.9
-cachetools==4.2.2
-cchardet==2.1.7
-certifi==2021.5.30
-cffi==1.14.5
-chardet==3.0.4
-charset-normalizer==2.0.7
-clarifai==2.6.2
-clean-fid==0.1.32
-click==8.1.3
-clip-anytorch==2.5.0
-cloudpickle==2.1.0
-configparser==3.8.1
-contourpy==1.0.5
-cryptography==3.4.7
-cycler==0.10.0
-Cython==0.29.30
-dask==2022.7.0
-datasets==2.6.1
-dcarte==0.3.38
-debugpy==1.5.1
-decorator==5.1.0
-deepdiff==5.8.1
-deepspeed==0.8.2
-defusedxml==0.7.1
-descartes==1.1.0
-dill==0.3.5.1
-distributed==2022.7.0
-docker-pycreds==0.4.0
-EasyProcess==0.3
-efficientnet==1.0.0
-einops==0.4.0
-emoji==1.2.0
-entrypoints==0.3
-et-xmlfile==1.1.0
-etils==0.6.0
-filelock==3.7.1
-flatbuffers==2.0.7
-fonttools==4.28.1
-frozenlist==1.3.1
-fsspec==2022.5.0
-ftfy==6.1.1
-future==0.18.2
+async-timeout==4.0.3
+attrs==23.2.0
+cachetools==5.4.0
+certifi==2024.7.4
+charset-normalizer==3.3.2
+contourpy==1.2.1
+cycler==0.12.1
+datasets==2.20.0
+dill==0.3.8
+dtaidistance==2.3.12
+filelock==3.15.4
+flatbuffers==24.3.25
+fonttools==4.53.1
+frozenlist==1.4.1
+fsspec==2024.5.0
 gast==0.4.0
-gdown==4.7.1
-gitdb==4.0.9
-GitPython==3.1.27
-google-api-core==2.10.1
-google-api-python-client==2.64.0
-google-auth==2.12.0
-google-auth-httplib2==0.1.0
-google-auth-oauthlib==0.4.4
+google-auth==2.32.0
+google-auth-oauthlib==0.4.6
 google-pasta==0.2.0
-googleapis-common-protos==1.56.4
-GPUtil==1.4.0
-grpcio==1.38.1
-h5py==3.1.0
-hdbscan==0.8.28
-HeapDict==1.0.1
-heartpy==1.2.7
-hjson==3.1.0
-httplib2==0.20.4
-huggingface-hub==0.14.1
-idna==2.8
-image-classifiers==1.0.0
-imageio==2.9.0
-imbalanced-learn==0.9.1
-importlib-resources==5.8.0
-ipykernel==6.4.2
-ipython==7.28.0
-ipython-genutils==0.2.0
-ipywidgets==7.6.5
-jedi==0.18.0
-Jinja2==3.0.2
-joblib==1.0.1
-jsonmerge==1.8.0
-jsonschema==2.6.0
-jupyter==1.0.0
-jupyter-client==7.0.6
-jupyter-console==6.4.0
-jupyter-core==4.9.0
-jupyter-http-over-ws==0.0.8
-jupyterlab-pygments==0.1.2
-jupyterlab-widgets==1.0.2
-k-diffusion==0.0.14
+grpcio==1.65.1
+h5py==3.11.0
+huggingface-hub==0.24.2
+idna==3.7
+Jinja2==3.1.4
+joblib==1.4.2
 keras==2.11.0
 Keras-Applications==1.0.8
 Keras-Preprocessing==1.1.2
-kiwisolver==1.3.1
-kornia==0.6.7
-lazy_loader==0.1rc2
-libclang==13.0.0
-littleutils==0.2.2
-llvmlite==0.38.1
-locket==1.0.0
-Markdown==3.3.4
-MarkupSafe==2.0.1
-matplotlib==3.6.1
-matplotlib-inline==0.1.3
-matrixprofile==1.1.10
-MeaningCloud-python==2.0.0
-minder.research-portal-client==0.1.3
-mistune==0.8.4
-mizani==0.7.3
-mne==0.23.0
-msgpack==1.0.4
-multidict==6.0.2
-multiprocess==0.70.13
-nbclient==0.5.4
-nbconvert==6.2.0
-nbformat==5.1.3
-nest-asyncio==1.5.1
-networkx==2.8.7
-ninja==1.11.1
-notebook==6.4.5
-numba==0.55.2
-numpy==1.21.4
-oauth2client==4.1.3
-oauthlib==3.1.1
-opencv-python==4.5.2.54
-openpyxl==3.1.2
+kiwisolver==1.4.5
+libclang==18.1.1
+llvmlite==0.43.0
+Markdown==3.6
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.9.1
+mdurl==0.1.2
+ml-dtypes==0.4.0
+mpmath==1.3.0
+multidict==6.0.5
+multiprocess==0.70.16
+namex==0.0.8
+networkx==3.3
+numba==0.60.0
+numpy==1.26.4
+nvidia-cublas-cu12==12.1.3.1
+nvidia-cuda-cupti-cu12==12.1.105
+nvidia-cuda-nvrtc-cu12==12.1.105
+nvidia-cuda-runtime-cu12==12.1.105
+nvidia-cudnn-cu12==9.1.0.70
+nvidia-cufft-cu12==11.0.2.54
+nvidia-curand-cu12==10.3.2.106
+nvidia-cusolver-cu12==11.4.5.107
+nvidia-cusparse-cu12==12.1.0.106
+nvidia-nccl-cu12==2.20.5
+nvidia-nvjitlink-cu12==12.5.82
+nvidia-nvtx-cu12==12.1.105
+oauthlib==3.2.2
+opencv-python==4.10.0.84
+opencv-python-headless==4.10.0.84
 opt-einsum==3.3.0
-ordered-set==4.1.0
-outdated==0.2.1
-packaging==21.3
-palettable==3.3.0
-pandas==1.5.0
-pandas-flavor==0.3.0
-pandocfilters==1.5.0
-parso==0.8.2
-partd==1.2.0
-pathlib==1.0.1
-pathtools==0.1.2
-patsy==0.5.2
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==9.2.0
-pingouin==0.5.1
-pkg_resources==0.0.0
-plotnine==0.8.0
-plyer==2.0.0
-portpicker==1.2.0
-prometheus-client==0.11.0
-promise==2.3
-prompt-toolkit==3.0.31
+optree==0.12.1
+packaging==24.1
+pandas==2.2.2
+pillow==10.4.0
 protobuf==3.19.6
-psutil==5.9.1
-ptyprocess==0.7.0
-py-cpuinfo==9.0.0
-py4j==0.10.9.2
-pyarrow==9.0.0
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pycares==4.2.2
-pycparser==2.20
-pydantic==1.10.5
-PyDrive==1.3.1
-pyeeg @ git+https://github.com/forrestbao/pyeeg.git@a6c18bb093e4748f9d9c208535a6ae024a0802b8
-Pygments==2.10.0
-pyOpenSSL==20.0.1
-pyparsing==2.4.7
-PyQt5==5.15.4
-PyQt5-Qt5==5.15.2
-PyQt5-sip==12.9.0
-PySimpleGUI==4.45.0
-PySocks==1.7.1
-pyspark==3.2.0
-python-dateutil==2.8.2
-python-dotenv==1.0.0
-python-telegram-bot==13.7
-python-vlc==3.0.12118
-pyts==0.12.0
-pytz==2021.1
-pytz-deprecation-shim==0.1.0.post0
-PyVirtualDisplay==2.2
-PyWavelets==1.1.1
-PyYAML==6.0
-pyzmq==22.3.0
-qtconsole==5.1.1
-QtPy==1.11.2
-regex==2021.7.1
-requests==2.25.1
-requests-oauthlib==1.3.0
-resize-right==0.0.2
-responses==0.10.15
-rsa==4.7.2
-schedule==1.1.0
-scikit-image==0.18.3
-scikit-learn==1.1.2
-scikit-multilearn==0.2.0
-scipy==1.7.3
-seaborn==0.11.1
-segmentation-models==1.0.1
-selenium==3.141.0
-semantic-version==2.8.5
-Send2Trash==1.8.0
-sensortransformer==0.1.6
-sentry-sdk==1.9.0
-setproctitle==1.3.2
-setuptools-rust==0.12.1
-setuptools-scm==6.3.2
-shortuuid==1.0.9
-sidetable==0.9.1
-simplegeneric==0.8.1
+pyarrow==17.0.0
+pyarrow-hotfix==0.6
+pyasn1==0.6.0
+pyasn1_modules==0.4.0
+Pygments==2.18.0
+pyparsing==3.1.2
+python-dateutil==2.9.0.post0
+pyts==0.13.0
+pytz==2024.1
+PyYAML==6.0.1
+requests==2.32.3
+requests-oauthlib==2.0.0
+rich==13.7.1
+rsa==4.9
+scikit-learn==1.5.1
+scipy==1.14.0
+seaborn==0.13.2
 six==1.16.0
-sklearn==0.0
-smmap==5.0.0
-sortedcontainers==2.4.0
-soupsieve==2.2.1
-statsmodels==0.13.2
-stumpy==1.11.1
-tabulate==0.8.9
-tblib==1.7.0
-tensorboard==2.11.0
+sympy==1.13.1
+tensorboard==2.11.2
 tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.0
+tensorboard-plugin-wit==1.8.1
 tensorflow==2.11.0
-tensorflow-addons==0.14.0
-tensorflow-datasets==4.6.0
+tensorflow-addons==0.23.0
 tensorflow-estimator==2.11.0
-tensorflow-gpu==2.7.0
-tensorflow-io-gcs-filesystem==0.23.1
-tensorflow-metadata==1.9.0
-termcolor==1.1.0
-terminado==0.12.1
-testpath==0.5.0
-tfgarden @ git+https://github.com/Shakshi3104/tfgarden.git@2b4c6321b4c94b7e69a17eb20a25d0f8c2cb92e5
-threadpoolctl==2.1.0
-tifffile==2022.8.12
-toml==0.10.2
-tomli==1.2.2
-toolz==0.12.0
-torch==1.12.1+cu113
-torchdiffeq==0.2.3
-torchsde==0.2.5
-torchvision==0.13.1+cu113
-tornado==6.2
-tqdm==4.64.1
-traitlets==5.1.1
-trampoline==0.1.2
-tsfresh==0.19.0
-typeguard==2.13.0
-typing_extensions==4.5.0
-tzdata==2021.5
-tzlocal==2.1
-uritemplate==4.1.1
-urllib3==1.24.3
-venv-pack==0.2.0
-wandb==0.14.2
-wcwidth==0.2.5
-webdriver-manager==3.8.5
-webdriverdownloader==1.1.0.3
-webencodings==0.5.1
-Werkzeug==2.0.1
-widgetsnbextension==3.5.1
-wrapt==1.12.1
-xarray==2022.3.0
-xgboost==1.4.2
-xlrd==2.0.1
-xxhash==3.0.0
-yarl==1.8.1
-yellowbrick==1.4
-zict==2.2.0
-zipp==3.8.1
+tensorflow-io-gcs-filesystem==0.37.1
+termcolor==2.4.0
+threadpoolctl==3.5.0
+torch==2.4.0
+torchvision==0.19.0
+tqdm==4.66.4
+triton==3.0.0
+typeguard==2.13.3
+typing_extensions==4.12.2
+tzdata==2024.1
+urllib3==2.2.2
+Werkzeug==3.0.3
+wrapt==1.16.0
+xxhash==3.4.1
+yarl==1.9.4
