+++ tgen/reconstruct_tseries.py
@@ -19,11 +19,11 @@ import tgen.calculate_ts_errors as err
 import time
 import seaborn as sns
 from sklearn.model_selection import StratifiedGroupKFold, GroupKFold
-import tgen.REC as rpts
+import tgen.recurrence_plots as rpts
 import tgen.GAF as gaf
-import tgen.GAF as gaf2
+#import tgen.GAF2 as gaf2
 import time
-
+import tgen.MTF as mtf2
 import os    
 
 """ Generates and saves a time series from a RP saving keeping also 
@@ -187,7 +187,85 @@ def generate_and_save_time_series_fromGAF(fold, dataset_folder, training_data, y
     
     #print("Numero total desplazamientos",t)
     return X_all_rec
+def generate_and_save_time_series_fromMTF(fold, dataset_folder, training_data, y_data, sj_train,dictionary, TIME_STEPS=129, data_type="train", single_axis=False, FOLDS_N=3, sampling="loso"):
 
+    subject_samples = 0
+    p_bar = tqdm(range(len(training_data)))
+    X_all_rec=[]
+    tiempos=[]
+    start_gl=time.time()
+    errores= []
+    #Colocar timer
+    
+    
+    for i in p_bar:
+      #start=time.time()
+      w = training_data[i]
+      sj = sj_train[i][0]
+      w_y = y_data[i]
+      w_y_no_cat = np.argmax(w_y)
+      #print("w_y", w_y, "w_y_no_cat", w_y_no_cat)
+      #print("w", w.shape)
+      error=[]
+      # Update Progress Bar after a while
+      #time.sleep(0.01)
+      #p_bar.set_description(f'[{data_type} | FOLD {fold} | Class {w_y_no_cat}] Subject {sj}')
+      path=f"{dataset_folder}plots/MTF/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/{w_y_no_cat}/"
+      path=f"{path}{sj}x{subject_samples}mtf.png"  
+      imagen = cv2.imread(path)  
+      imagen = cv2.cvtColor(imagen, cv2.COLOR_BGR2RGB)
+      ##We need to change path 
+      mtf=mtf2.Reconstruct_MTF(imagen,dictionary,subject_samples)
+      #end=time.time()
+      
+      #Si queremos guardar los datos
+      path=f"{dataset_folder}tseries/MTF/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/{w_y_no_cat}/{sj}x{subject_samples}.npy"
+      np.save(path, np.array(mtf))
+      
+      #print(rp)
+      X_all_rec.append(mtf)
+      #print("X_ALL",np.array(X_all_rec))
+      #tiempo=end-start
+      #tiempos.append(tiempo)
+      #Guardar la rp en el path indicado con un nombre adecuado
+      #print(tiempos)
+      #w=w.reshape(3,129)
+      #w=w[:, 0]
+      #w=w[1:]
+      #print("Forma del RP original y calculada",w.shape,rp.shape)
+      
+      #print("normal",w,"calculada",rp)
+      
+      error_abs,error_r,error_q,error_std,error_p,te=err.ts_error(w,mtf)
+      
+      error.append(error_abs)
+      error.append(error_r)
+      error.append(error_q)
+      error.append(error_std)
+      error.append(error_p)
+      errores.append(error)
+      #print("errores",np.array(error))
+      #print("errores",np.array(errores))
+      
+      subject_samples += 1
+
+    end_gl=time.time()
+    ttotal=end_gl-start_gl
+
+    #maybe we should calculate here the global of errors and the mean.
+    #maybe we should calculate here the global of errors and the mean.
+    archivoX_all=f"{dataset_folder}tseries/MTF/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/X_all_mtf.npy"
+    np.save(archivoX_all,np.array(X_all_rec))
+    archivoerrores=f"{dataset_folder}tseries/MTF/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/errores_mtf.npy"
+    np.save(archivoerrores,np.array(errores))
+
+    #archivotiempos=f"{dataset_folder}tseries/recurrence_plot/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/tiempos_rec.npy"
+    #np.save(archivotiempos,np.array(ttotal/subject_samples))
+    #print("Tiempo medio",np.mean(tiempos))
+    print("Tiempo medio total",np.mean(ttotal/subject_samples))
+    
+    #print("Numero total desplazamientos",t)
+    return X_all_rec
 #Generates all time series
 def generate_all_time_series(X_train, y_train, sj_train, dataset_folder="/home/adriano/Escritorio/TFG/data/WISDM/", TIME_STEPS=129,  FOLDS_N=3, sampling="loso",reconstruction="all"):
   groups = sj_train 
@@ -246,12 +324,27 @@ def generate_all_time_series(X_train, y_train, sj_train, dataset_folder="/home/a
     
 
     ##aqui añadir una opcion si quieres todas o una en especifico
-    if reconstruction=="all":
-      #generate_and_save_time_series_fromRP(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
-      generate_and_save_time_series_fromGAF(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
-      #print(train_index[0:6])
+    if reconstruction=="MTF":
+      generate_and_save_time_series_fromMTF(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+      
       #arch_training_data=f"{dataset_folder}tseries/recurrence_plot/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/train/training_data.npy"
       #np.save(arch_training_data,np.array(training_data))
       print("se hace 1 vez")
       break
-      #generate_and_save_time_series_fromRP(fold, dataset_folder, validation_data, y_validation_data, sj_validation_data, TIME_STEPS=TIME_STEPS, data_type="test", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+    if reconstruction=="GAF":
+      generate_and_save_time_series_fromGAF(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+      print("se hace 1 vez")
+      break
+    if reconstruction=="REC":
+      generate_and_save_time_series_fromRP(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+
+      print("se hace 1 vez")
+      break
+    if reconstruction=="all":
+      generate_and_save_time_series_fromRP(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+      generate_and_save_time_series_fromGAF(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+      generate_and_save_time_series_fromMTF(fold, dataset_folder, training_data, y_training_data, sj_training_data,dictionary,TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
+  
+      print("se hace 1 vez")
+      break
+
