+++ tgen/recurrence_plots.py
@@ -1,17 +1,13 @@
-import os
-import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.pylab import rcParams
 import math
 # from tqdm.notebook import tqdm, trange #progress bars for jupyter notebook
 from tqdm.auto import trange, tqdm #progress bars for pyhton files (not jupyter notebook)
-
-import time
-import seaborn as sns
 from sklearn.model_selection import StratifiedGroupKFold, GroupKFold
-
+from sklearn.manifold import MDS
 from PIL import Image
+from scipy.sparse.csgraph import dijkstra
 def NormalizeMatrix_Adri(_r):
     dimR = _r.shape[0]
     _max=66.615074
@@ -418,182 +414,161 @@ def SaveRP_fran(x, axis, sj=0, item_idx=0, action=None, normalized=True, path=No
         # plt.close('all')
     return newImage
 
-# def recover_time_serie_from_varRP(varRP):
-#     dimR = len(varRP)
+def RecurrrenceTreshold(rp,X):
+    distances=np.abs(rp)
+    epsilon_X=np.percentile(distances,X)
+    return epsilon_X
 
-#     #R = np.zeros((dimR,dimR))
-#     R = np.eye(dimR)
-#     for i in range(dimR):
-#         for j in range(dimR):
-#             if Cosin2vec(list(map(lambda x:x[0]-x[1], zip(s[i], s[j]))), [1,1]) >= pow(2, 0.5)/2:
-#                 sign =1.0
-#             else:
-#                 sign =-1.0
-#             R[i][j] = sign*Distance2dim(s[i],s[j])
-#     return R
+def FuncionC(rp,cord1,cord2,method="Euclid"):
+    if(method=="Euclid"):
+       THRESHOLD=RecurrrenceTreshold(rp,75) # THRESHOLD=RecurrrenceTreshold(rp,75)
+       valor=0
+       d=rp[cord1][cord2]
+       if((d<=THRESHOLD)and(d>=(-THRESHOLD))):
+          valor=1
+    return valor  
+def CreateCostMatrix(rp) :
+    CostM=np.zeros_like(rp)
+    N,M=CostM.shape
+    for i in range(N):
+         for j in range(M):
+             CostM[i][j]=FuncionC(rp,i,j,"Euclid")
+    return CostM
 
-#@title plot  reconstruction
 
-def plot_reconstruct_time_series( df, activity, dataset_folder="/home/fmgarmor/proyectos/TGEN-timeseries-generation/data/WISDM/", TIME_STEPS=129, axes_cols=['x_axis', 'y_axis', 'z_axis'], signal_col="signal", timestep_start_idx=0, subject=0, grid=True, figsize=(10,4), saveFig=False, showPlot=False):
-  sns.set_style("darkgrid")
+def calculate_shortest_path_matrix(Wg):
+    shortest_path_matrix = dijkstra(Wg, directed=True)
+    return shortest_path_matrix + 1e-10
 
-  columns = axes_cols.copy()
-  columns.append(signal_col)
-  data = df[columns]
-  # print(data)
-  
-  if len(data) > 0:
-    # print("Plots")
-    # fig, ax = plt.subplots(3,1)
-    # for ax in ax.flatten():
-    #   data.set_index('signal',append=True).unstack().plot(ax=ax,subplots=True, figsize=figsize, title=activity, xticks=range(0, TIME_STEPS+1, TIME_STEPS // 2))
-    df_list = []
-    for y in axes_cols:
-      # print(y)
-      # print( len(axes_cols) >1)
-      d = data.set_index(signal_col,append=True).unstack()[y] #if len(axes_cols) >1 else data
-      df_list.append(d)
-      # print(d)
-      # axis = d.plot(figsize=figsize, title=activity, xticks=range(0, TIME_STEPS+1, TIME_STEPS // 2))
-      # for ax in axis:
-      # axis.legend(loc='lower left', bbox_to_anchor=(1.0, 0.5))
-      # axis.grid(grid)
 
-    # make a list of all dataframes 
-    nrows=len(axes_cols)
 
-    fig, axes = plt.subplots(nrows=nrows, ncols=1, sharex=True) #, sharey=True)
-    # print("axes.shape", axes.shape)
-    # plot counter
-    count=0
-    for r in range(nrows):
+   
+def weigthed_graphRP(rp):
+    CostM=CreateCostMatrix(rp)
+    weighted_adjacency_matrix = np.zeros_like(rp) 
+    nonzero_indices = np.nonzero(CostM)
 
-      df_list[count].plot(ax=axes[r,], figsize=figsize, xticks=range(0, len(df) // 2 +1, (len(df) // 2) // 4))
-      axes[r,].legend(loc='lower left', bbox_to_anchor=(1.0, 0.5))
-      axes[r,].grid(grid)
-      axes[r,].set_ylabel(axes_cols[r])
-      count+=1
+    for i, j in zip(*nonzero_indices):
+        Gi = np.nonzero(CostM[i])[0]
+        Gj = np.nonzero(CostM[j])[0]
+        intersection = len(np.intersect1d(Gi, Gj))
+        union = len(np.union1d(Gi, Gj))
+        weighted_adjacency_matrix[i, j] = 1 - (intersection / union)
+                
+    return weighted_adjacency_matrix
     
-    plt.xlabel("Timestep")
-    fig.suptitle(activity)
+def reconstruct_time_series(shortest_path_matrix, ep=0.0, small_constant=1e-10):
+    # Forzar la simetría en la matriz
+    symmetric_shortest_path_matrix = 0.5 * (shortest_path_matrix + shortest_path_matrix.T)
 
+    # Reemplazar los infinitos en la matriz con un valor grande pero finito
+    finite_shortest_path_matrix = np.where(
+        np.isfinite(symmetric_shortest_path_matrix),
+        symmetric_shortest_path_matrix,
+        ep  # Ajusta este valor según sea necesario
+    )
 
-    # axis = data.plot(subplots=True, figsize=figsize, title=activity, xticks=range(0, TIME_STEPS+1, TIME_STEPS // 2))
-    # for ax in axis:
-    #     ax.legend(loc='lower left', bbox_to_anchor=(1.0, 0.5))
-    #     ax.grid(grid)
+    #print("mfinita",finite_shortest_path_matrix)
 
-    save_path = f"{dataset_folder}visualization/{activity}/time-series-reconstruction/{subject}/"
-    os.makedirs(save_path, exist_ok=True)
-    # fig = axis.get_figure()
-    # plt.tight_layout()
+     # Add a small constant to avoid division by zero
+    finite_shortest_path_matrix += small_constant    
+
+    # Apply MDS to get a 2D representation of the shortest path matrix
+    mds = MDS(n_components=2, dissimilarity='precomputed',random_state=1)
+    embedded_coords = mds.fit_transform(finite_shortest_path_matrix)
     
-    if saveFig:
-      # print(f"{save_path}visual_cl_{activity}_timestep_{TIME_STEPS}_w_{timestep_start_idx}.png")
-      fig.savefig(f"{save_path}visual_cl_{activity}_timestep_{TIME_STEPS}_w_{timestep_start_idx}.png", dpi=300, bbox_inches='tight')
+    #----
+    # Calcular la matriz de covarianza
+    cov_matrix = np.cov(embedded_coords, rowvar=False)
+    #print("mds",embedded_coords)
+    # Calcular los valores y vectores propios de la matriz de covarianza
+    eigenvalues, eigenvectors = np.linalg.eig(cov_matrix)
 
-    if showPlot:
-      plt.show()
-    plt.close('all')
+    # Encontrar el índice del valor propio más grande
+    max_eigenvalue_index = np.argmax(eigenvalues)
 
-  else:
-    print("no data to plot")
+    # Seleccionar la columna correspondiente al valor propio más grande en la matriz embedded_coords
+    selected_column = embedded_coords[:, max_eigenvalue_index][:, np.newaxis]
 
-def generate_and_save_recurrence_plot(fold, dataset_folder, training_data, y_data, sj_train, TIME_STEPS=129, data_type="train", single_axis=False, FOLDS_N=3, sampling="loso"):
-    subject_samples = 0
-    p_bar = tqdm(range(len(training_data)))
-
-    for i in p_bar:
-      w = training_data[i]
-      sj = sj_train[i][0]
-      w_y = y_data[i]
-      w_y_no_cat = np.argmax(w_y)
-      print("w_y", w_y, "w_y_no_cat", w_y_no_cat)
-      print("w", w.shape)
-
-      # Update Progress Bar after a while
-      time.sleep(0.01)
-      p_bar.set_description(f'[{data_type} | FOLD {fold} | Class {w_y_no_cat}] Subject {sj}')
+    return selected_column
+def Reconstruct_RP(img,dictionary,valor):
+    _r= img[:,:,0].astype('float')
+    _g= img[:,:,1].astype('float')
+    _b= img[:,:,2].astype('float')
+    #Obtengo cada una de las recurrence plots
     
-      # #-------------------------------------------------------------------------
-      # # only for degugging in notebook
-      # #-------------------------------------------------------------------------
-      # df_original_plot = pd.DataFrame(w, columns=["x_axis", "y_axis", "z_axis"])
-      # df_original_plot["signal"] = np.repeat("Original", df_original_plot.shape[0])
-      # df_original_plot = df_original_plot.iloc[:-1,:]
-      # plot_reconstruct_time_series(df_original_plot, "Walking", subject=sj)
-      # #-------------------------------------------------------------------------
-
-      #print(f"{'*'*20}\nSubject: {sj} (window: {i+1}/{len(training_data)} | label={y})\n{'*'*20}")
-      #print("Window shape",w.shape)
-      if fold < 0:
-        img = SavevarRP_XYZ(w, sj, subject_samples, "x", normalized = 1, path=f"{dataset_folder}plots/recurrence_plot/sampling_{sampling}/{data_type}/{w_y_no_cat}/", TIME_STEPS=TIME_STEPS)
-      else:
-        img = SavevarRP_XYZ(w, sj, subject_samples, "x", normalized = 1, path=f"{dataset_folder}plots/recurrence_plot/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/{w_y_no_cat}/", TIME_STEPS=TIME_STEPS)
-      print("w image (RP) shape:", np.array(img).shape)
-      
-      if single_axis and img is not None:
-        #also, save each single data column values
-        for col in range(w.shape[1]):
-          if fold < 0:
-            img = SavevarRP_fran(w, sj, subject_samples, "x", normalized = 1, path=f"{dataset_folder}plots/recurrence_plot/sampling_{sampling}/{data_type}/{w_y_no_cat}/", TIME_STEPS=TIME_STEPS)
-          else:
-            img = SavevarRP_fran(w, col, sj, subject_samples, "x", normalized = 1, path=f"{dataset_folder}plots/single_axis/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/{data_type}/{w_y_no_cat}/", TIME_STEPS=TIME_STEPS)
-          print("w single axis (RP) shape:", img.shape)
-      subject_samples += 1
-      
+    #print("X2",_r[1][4])
+    ##PREGUNTAR SI ES LO MISMO O HABRIA QUE PASAR de 255 a [0-1] y posteriormente a min max
+    _r=np.interp(_r,(0,255),(dictionary[valor][1][0],dictionary[valor][0][0]))
+    _g=np.interp(_g,(0,255),(dictionary[valor][1][1],dictionary[valor][0][1]))
+    _b=np.interp(_b,(0,255),(dictionary[valor][1][2],dictionary[valor][0][2]))
+    
+    
+    R= []
+    R.append(_r)
+    R.append(_g)
+    R.append(_b)
+    n=len(R)
+    N=[]
+    ninversiones=0
+    for i in range(0,n):
+        wg=weigthed_graphRP(R[i]) 
+        spm=calculate_shortest_path_matrix(wg)  
+        ##we multiply * -1 ya que se invierte al calcular los valores propios
+        rp=reconstruct_time_series(spm, ep=0.0)
+        rp,a=fix_rotationscale(R[i],rp,i,dictionary[valor])
+        ninversiones=ninversiones+a
+        N.append(rp)
+    N=np.array(N)
+    return N,ninversiones
+def fix_rotationscale(rporiginal,seriereconstruida,i,dictionary,TIME_STEPS=129):
+     #print(dictionary)
+     MAX=dictionary[0]
+     MIN=dictionary[1]
+     
+     #escalo la serie para que tenga los valores 
+     
+     
+     #n=(seriereconstruida-min_val)/(max_val-min_val)
+     #print(seriereconstruida.shape)
+     #Scaling part
+     #n=seriereconstruida
+     n=np.append(seriereconstruida,np.mean(seriereconstruida))
+     
+     #posi= MIN[i]+n*(MAX[i]-1)
+     min_a, max_a = np.min(n), np.max(n)  
      
 
+    # Transformar los puntos
+     #posi =min_b + n * (max_b - min_b)
+     posi=np.interp(n,(min_a,max_a),(MIN[i], MAX[i]))
+     rposi=varRP2(posi, TIME_STEPS)    
+     
+     n=n*-1+1  
+     min_a, max_a = np.min(n), np.max(n)  
+     
+     #nega = min_b + n * (max_b - min_b)
+     nega=np.interp(n,(min_a,max_a),(MIN[i], MAX[i]))
+     rnega=varRP2(nega, TIME_STEPS)
+     #print(n)
+     #print(nega)
 
-def generate_all_recurrence_plots(X_train, y_train, sj_train, dataset_folder="/home/fmgarmor/proyectos/TGEN-timeseries-generation/data/WISDM/", TIME_STEPS=129,  FOLDS_N=3, sampling="loso"):
-  groups = sj_train 
-  if sampling == "loto":
-    #TODO change 100 for an automatic extracted number greater than the max subject ID: max(sj_train)*10
-    groups = [[int(sj[0])+i*100+np.argmax(y_train[i])+1] for i,sj in enumerate(sj_train)]
-
-  # if DATASET_NAME == "WISDM": #since wisdm is quite balanced
-  sgkf = StratifiedGroupKFold(n_splits=FOLDS_N)
-  # elif DATASET_NAME == "MINDER" or DATASET_NAME == "ORIGINAL_WISDM": 
-  #   sgkf = StratifiedGroupKFold(n_splits=FOLDS_N)
-
-  accs = []
-  y_train_no_cat = [np.argmax(y) for y in y_train]
-  p_bar_classes = tqdm(range(len(np.unique(y_train_no_cat))))
-  all_classes = np.unique(y_train_no_cat)
-  print("Classes available: ", all_classes)
-  for fold in range(FOLDS_N):
-    for i in p_bar_classes:
-        y = all_classes[i]
-        time.sleep(0.01) # Update Progress Bar after a while
-        os.makedirs(f"{dataset_folder}plots/recurrence_plot/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/train/{y}/", exist_ok=True) 
-        os.makedirs(f"{dataset_folder}plots/recurrence_plot/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/test/{y}/", exist_ok=True) 
-        os.makedirs(f"{dataset_folder}plots/single_axis/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/train/{y}/", exist_ok=True)
-        os.makedirs(f"{dataset_folder}plots/single_axis/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/test/{y}/", exist_ok=True)
-
-  for fold, (train_index, val_index) in enumerate(sgkf.split(X_train, y_train_no_cat, groups=groups)):
-
-    # if fold != 2:
-    #   continue
-
-    # print(f"{'*'*20}\nFold: {fold}\n{'*'*20}")
-    # print("Train index", train_index)
-    # print("Validation index", val_index)
-    training_data = X_train[train_index,:,:]
-    validation_data = X_train[val_index,:,:]
-    y_training_data = y_train[train_index]
-    y_validation_data = y_train[val_index]
-    sj_training_data = sj_train[train_index]
-    sj_validation_data = sj_train[val_index]
-
-    print("training_data.shape", training_data.shape, "y_training_data.shape", y_training_data.shape, "sj_training_data.shape", sj_training_data.shape)
-    print("validation_data.shape", validation_data.shape, "y_validation_data.shape", y_validation_data.shape, "sj_validation_data.shape", sj_validation_data.shape)
-
-    #arch_training_data=f"{dataset_folder}tseries/recurrence_plot/sampling_{sampling}/{FOLDS_N}-fold/fold-{fold}/train/training_data.npy"
-    #np.save(arch_training_data,np.array(train_index))
-   
-    generate_and_save_recurrence_plot(fold, dataset_folder, training_data, y_training_data, sj_training_data, TIME_STEPS=TIME_STEPS, data_type="train", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
-    generate_and_save_recurrence_plot(fold, dataset_folder, validation_data, y_validation_data, sj_validation_data, TIME_STEPS=TIME_STEPS, data_type="test", single_axis=False, FOLDS_N=FOLDS_N, sampling=sampling)
-    
-
+     rp=[]
+     #print(rporiginal.shape,rposi.shape)
+     
+     #error_absolutoa, error_relativoa= calcular_errores(rporiginal[:60], rposi[0])
+     #error_absolutob, error_relativob= calcular_errores(rporiginal[:60], rnega[0])
+     distancia_euclidianaa = np.linalg.norm(rporiginal - rposi)
+     distancia_euclidianab = np.linalg.norm(rporiginal - rnega)
+     inverted=0
+     if distancia_euclidianab<distancia_euclidianaa :
+         rp=nega[:128]
+         inverted+=1
+     else :
+        rp=posi[:128]
+     
+        #print("posi")
+        
+     return  rp,inverted
